fix snapshots and heading name
