+++ src/components/Title/__stories__/Title.stories.mdx
@@ -2,10 +2,10 @@ import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
 import { createStoryMetaSettingsDecorator, Tip, AlphaWarning } from "../../../storybook";
 import Flex from "../../Flex/Flex";
 import { createComponentTemplate } from "vibe-storybook-components";
-import Heading from "../Heading";
-import { OVERFLOW_TITLE_CONTAINER_ID, ONE_LINE_ELLIPSIS_TEST_ID } from "../__tests__/heading-tests-constants";
-import styles from "./Heading.stories.module.scss";
-import { titleOverflowSuite } from "../__tests__/Heading.interactions";
+import Heading from "../Title";
+import { OVERFLOW_TITLE_CONTAINER_ID, ONE_LINE_ELLIPSIS_TEST_ID } from "../__tests__/title-tests-constants";
+import styles from "./Title.stories.module.scss";
+import { titleOverflowSuite } from "../__tests__/Title.interactions";
 
 export const metaSettings = createStoryMetaSettingsDecorator({
   component: Heading,
