+++ src/components/Title/__tests__/Title.interactions.ts
@@ -1,6 +1,6 @@
 import { testHoverTooltipTrigger } from "../../Tooltip/__tests__/tooltip.interactions";
 import { Screen } from "@testing-library/react";
-import { ONE_LINE_ELLIPSIS_TEST_ID } from "./heading-tests-constants";
+import { ONE_LINE_ELLIPSIS_TEST_ID } from "./title-tests-constants";
 
 import { interactionSuite } from "../../../tests/interactions-utils";
